add: share -> share_helperに変更・utils/helperにファイルを移動しました
